refac: Move registration function into app.templates
